Rename: JWT 관련 폴더 이동
